refactor(cli): extract parser commands enum
